users and auth WIP

+++ src/main/java/com/example/application/data/repository/CourseRepository.java
@@ -3,4 +3,4 @@ package com.example.application.data.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.application.data.entity.Course;
 
-public interface CourseRepository extends JpaRepository<Course, Integer> {}
+public interface CourseRepository extends JpaRepository<Course, Long> {}
